Store links properly. One marker per link

+++ app/models/network.js
@@ -627,6 +627,8 @@ module.exports = class Network {
     return database.query(`
       SELECT
         plan_links.id AS id,
+        plan_links.from_link_id,
+        plan_links.to_link_id,
         ST_AsGeoJSON(nn1.geom)::json AS from_geom,
         ST_AsGeoJSON(nn2.geom)::json AS to_geom
       FROM client.plan_links
@@ -636,14 +638,15 @@ module.exports = class Network {
     `, [plan_id])
   }
 
-  static saveBackhaulLinks (plan_id, ids) {
+  static saveBackhaulLinks (plan_id, fromIds, toIds) {
+    if (fromIds.length !== toIds.length) {
+      return Promise.reject(new Error('fromIds and toIds should have the same length'))
+    }
     return database.execute('DELETE FROM client.plan_links WHERE client.plan_links.plan_id = $1', [plan_id])
       .then(() => {
-        if (ids.length < 2) return
-        var prev = ids[0]
-        return pync.series(ids.slice(1), (id) => {
-          var params = [prev, id, plan_id]
-          prev = id
+        var i = 0
+        return pync.series(fromIds, (id) => {
+          var params = [id, toIds[i++], plan_id]
           return database.execute('INSERT INTO client.plan_links (from_link_id, to_link_id, plan_id) VALUES ($1, $2, $3)', params)
         })
       })
